backend_python/requirements.txt
fastapi
uvicorn
pymongo
pydantic
python-dotenv
passlib[bcrypt]
pyjwt
python-multipart
